refactor(compiler): extract spec compiler into standalone package
Extract questfoundry-compiler as independent package to serve multiple
runtimes (static and dynamic compilation).

**New Package: lib/compiler/ (questfoundry-compiler)**
- Package name: questfoundry-compiler (PyPI) / questfoundry_compiler (import)
- Version: 0.1.0
- CLI: qf-compile (moved from questfoundry-py)

Added:
- Complete package infrastructure (pyproject.toml, README, LICENSE, .gitignore, agents.md)
- GitHub workflows (lint-test.yml, publish-compiler.yml)
- All compiler source files (~1,629 lines)
- All compiler tests (23 tests, 21 passing)

**Updated: lib/python/ (questfoundry-py)**
- Removed [project.scripts] qf-compile CLI entry point
- Added build dependency: questfoundry-compiler>=0.1.0,<1.0.0
- Updated bundle_resources.py to import from questfoundry_compiler
- Removed lib/python/src/questfoundry/compiler/ directory
- Removed lib/python/tests/compiler/ directory

**Benefits:**
- Single compiler serves both static (build-time) and dynamic (runtime) use cases
- No code duplication between L6a (agentic) and L6b (web) runtimes
- Independent versioning and release cycles
- Reduced published package size for questfoundry-py

**Migration Notes:**
- For development: Install compiler as editable dependency
- For production builds: Compiler fetched from PyPI as build dependency
- Breaking changes to compiler require coordinated version bumps

Related to gist: https://gist.github.com/pvliesdonk/087e12e9a1ba5587fcc16234ded0a6bc
