+++ lib/compiler/.github/workflows/publish-compiler.yml
@@ -1,0 +1,175 @@
+name: Publish Compiler
+
+on:
+  workflow_dispatch:
+    inputs:
+      destination:
+        description: "Publish to PyPi"
+        required: true
+        type: choice
+        options:
+          - pypi
+          - testpypi
+      bump_version:
+        description: "Automatically bump version and create a release tag"
+        required: false
+        type: boolean
+        default: true
+
+jobs:
+  prepare-release:
+    name: Prepare Release Inputs
+    runs-on: ubuntu-latest
+    permissions:
+      contents: write
+    outputs:
+      publish_ref: ${{ steps.bump.outputs.publish_ref || steps.manual-ref.outputs.publish_ref }}
+      new_tag: ${{ steps.bump.outputs.new_tag }}
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+          token: ${{ secrets.GITHUB_TOKEN }}
+
+      - name: Set up Python (release only)
+        if: inputs.destination == 'pypi' && inputs.bump_version == true
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.11"
+
+      - name: Install Commitizen
+        if: inputs.destination == 'pypi' && inputs.bump_version == true
+        run: pip install commitizen
+
+      - name: Configure Git user
+        if: inputs.destination == 'pypi' && inputs.bump_version == true
+        run: |
+          git config user.name "github-actions[bot]"
+          git config user.email "github-actions[bot]@users.noreply.github.com"
+
+      - name: Bump version, tag, and push
+        if: inputs.destination == 'pypi' && inputs.bump_version == true
+        id: bump
+        working-directory: lib/compiler
+        run: |
+          cz bump --yes
+          git push --follow-tags
+          NEW_TAG=$(git describe --tags --abbrev=0)
+          echo "new_tag=${NEW_TAG}" >> $GITHUB_OUTPUT
+          echo "publish_ref=${NEW_TAG}" >> $GITHUB_OUTPUT
+
+      - name: Use current SHA without bump
+        if: inputs.destination == 'testpypi' || (inputs.destination == 'pypi' && inputs.bump_version == false)
+        id: manual-ref
+        run: echo "publish_ref=${GITHUB_SHA}" >> $GITHUB_OUTPUT
+
+  build:
+    name: Build Distribution
+    needs: prepare-release
+    runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: lib/compiler
+    steps:
+      - name: Checkout source to publish ref
+        uses: actions/checkout@v4
+        with:
+          ref: ${{ needs.prepare-release.outputs.publish_ref }}
+
+      - name: Install uv
+        uses: astral-sh/setup-uv@v3
+        with:
+          version: "latest"
+
+      - name: Set up Python
+        run: uv python install 3.11
+
+      - name: Install dependencies
+        run: uv sync
+
+      - name: Build package
+        run: uv build
+
+      - name: Store the distribution packages
+        uses: actions/upload-artifact@v4
+        with:
+          name: compiler-package-distributions
+          path: lib/compiler/dist/
+
+  publish-to-pypi:
+    name: Publish Compiler to PyPI
+    if: inputs.destination == 'pypi'
+    needs:
+      - build
+      - prepare-release
+    runs-on: ubuntu-latest
+    environment:
+      name: pypi-compiler
+      url: https://pypi.org/p/questfoundry-compiler
+    permissions:
+      id-token: write
+
+    steps:
+      - name: Download distributions
+        uses: actions/download-artifact@v4
+        with:
+          name: compiler-package-distributions
+          path: dist/
+
+      - name: Publish to PyPI
+        uses: pypa/gh-action-pypi-publish@release/v1
+
+  publish-to-testpypi:
+    name: Publish Compiler to TestPyPI
+    if: inputs.destination == 'testpypi'
+    needs:
+      - build
+      - prepare-release
+    runs-on: ubuntu-latest
+    environment:
+      name: testpypi-compiler
+      url: https://test.pypi.org/p/questfoundry-compiler
+    permissions:
+      id-token: write
+
+    steps:
+      - name: Download distributions
+        uses: actions/download-artifact@v4
+        with:
+          name: compiler-package-distributions
+          path: dist/
+
+      - name: Publish to TestPyPI
+        uses: pypa/gh-action-pypi-publish@release/v1
+        with:
+          repository-url: https://test.pypi.org/legacy/
+
+  github-release:
+    name: Create GitHub Release
+    if: inputs.destination == 'pypi' && inputs.bump_version == true
+    needs:
+      - publish-to-pypi
+      - prepare-release
+    runs-on: ubuntu-latest
+    permissions:
+      contents: write
+      id-token: write
+
+    steps:
+      - name: Download distributions
+        uses: actions/download-artifact@v4
+        with:
+          name: compiler-package-distributions
+          path: dist/
+
+      - name: Create GitHub Release
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          NEW_TAG: ${{ needs.prepare-release.outputs.new_tag }}
+        run: |
+          gh release create "$NEW_TAG" \
+            --repo '${{ github.repository }}' \
+            --notes "Compiler release $NEW_TAG" \
+            dist/*
