+++ lib/compiler/pyproject.toml
@@ -1,21 +1,18 @@
 [build-system]
-requires = [
-    "hatchling",
-    "questfoundry-compiler>=0.1.0,<1.0.0",
-]
+requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
-name = "questfoundry-py"
+name = "questfoundry-compiler"
 
-version = "0.6.0"
+version = "0.1.0"
 
-description = "Python library for QuestFoundry protocol and artifact management"
+description = "Spec compiler for QuestFoundry - transforms atomic behavior primitives into runtime artifacts"
 readme = "README.md"
 license = {text = "MIT"}
 requires-python = ">=3.11"
 authors = [{name = "Peter Liesdonk", email = "peter@liesdonk.nl"} ]
-keywords = ["questfoundry", "protocol", "artifacts", "interactive-fiction", "narrative"]
+keywords = ["questfoundry", "compiler", "spec-compiler", "behavior-primitives", "interactive-fiction"]
 classifiers = [
     "Development Status :: 3 - Alpha",
     "Intended Audience :: Developers",
@@ -25,57 +22,38 @@ classifiers = [
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
     "Programming Language :: Python :: 3.14",
-    "Topic :: Software Development :: Libraries :: Python Modules",
+    "Topic :: Software Development :: Code Generators",
+    "Topic :: Software Development :: Compilers",
     "Typing :: Typed",
 ]
 
 dependencies = [
-    "pydantic>=2.0",
-    "jsonschema>=4.0",
-    "httpx>=0.25",
-    "python-dotenv>=1.0",
     "pyyaml>=6.0",
 ]
 
 [project.optional-dependencies]
-openai = ["openai>=1.0"]
-ollama = ["ollama>=0.1"]
-google = ["google-generativeai>=0.3"]
-all-providers = [
-    "openai>=1.0",
-    "ollama>=0.1",
-    "google-generativeai>=0.3",
-]
-
 dev = [
     "pytest>=7.0",
     "pytest-cov>=4.0",
     "ruff>=0.1.0",
     "mypy>=1.0",
     "pre-commit>=3.0",
-    "types-jsonschema>=4.0",
     "types-pyyaml>=6.0",
     "commitizen>=3.12.0",
 ]
 
-docs = [
-    "mkdocs>=1.5.0",
-    "mkdocs-material>=9.0.0",
-    "mkdocstrings[python]>=0.23.0",
-    "pymdown-extensions>=10.0",
-    "markdown>=3.4",
-]
-
 [project.urls]
-Homepage = "https://github.com/pvliesdonk/questfoundry-py"
+Homepage = "https://github.com/pvliesdonk/questfoundry"
 Documentation = "https://github.com/pvliesdonk/questfoundry-spec"
-Specification = "https://github.com/pvliesdonk/questfoundry-spec"
-Schemas = "https://questfoundry.liesdonk.nl/schemas/"
+Repository = "https://github.com/pvliesdonk/questfoundry"
+"Bug Tracker" = "https://github.com/pvliesdonk/questfoundry/issues"
+
+[project.scripts]
+qf-compile = "questfoundry_compiler.cli:main"
 
 [tool.hatch.envs.default.scripts]
-bundle = "python scripts/bundle_resources.py"
 test = "pytest {args:tests}"
-test-cov = "pytest --cov=questfoundry --cov-report=term-missing {args:tests}"
+test-cov = "pytest --cov=questfoundry_compiler --cov-report=term-missing {args:tests}"
 lint = "ruff check src tests"
 format = "ruff format src tests"
 format-check = "ruff format --check src tests"
@@ -83,7 +61,7 @@ typecheck = "mypy src"
 all-checks = ["format-check", "lint", "typecheck", "test"]
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/questfoundry"]
+packages = ["src/questfoundry_compiler"]
 
 [tool.hatch.build.targets.sdist]
 include = [
@@ -107,17 +85,12 @@ disallow_untyped_defs = true
 disable_error_code = ["misc", "no-any-return", "attr-defined"]
 
 [[tool.mypy.overrides]]
-module = ["pydantic", "jsonschema.*", "yaml.*", "openai.*", "ollama.*", "httpx.*", "google.*", "vertexai.*"]
+module = ["yaml.*"]
 ignore_missing_imports = true
-
-[[tool.mypy.overrides]]
-module = ["questfoundry.providers.text.gemini", "questfoundry.providers.image.imagen"]
-disable_error_code = ["arg-type"]
 
 [tool.ruff]
 line-length = 88
 target-version = "py311"
-exclude = ["spec"]
 src = ["src", "tests"]
 
 [tool.ruff.lint]
@@ -125,20 +98,12 @@ select = ["E", "F", "W", "I"]
 
 [tool.commitizen]
 name = "cz_conventional_commits"
-version = "0.6.0"
+version = "0.1.0"
 version_files = [
-    "src/questfoundry/version.py:__version__",
     "pyproject.toml:version"
 ]
 update_changelog_on_bump = true
 changelog_file = "CHANGELOG.md"
-tag_format = "v$version"
-bump_message = "bump: version $current_version → $new_version"
+tag_format = "compiler-v$version"
+bump_message = "bump(compiler): version $current_version → $new_version"
 annotated_tag = true
-
-[dependency-groups]
-dev = [
-    "pytest>=8.4.2",
-    "pytest-cov>=7.0.0",
-    "types-jsonschema>=4.25.1.20251009",
-]
