+++ lib/compiler/src/questfoundry_compiler/__init__.py
@@ -1,6 +1,6 @@
 """QuestFoundry Spec Compiler - Transform behavior primitives into runtime artifacts."""
 
-from questfoundry.compiler.spec_compiler import SpecCompiler
-from questfoundry.compiler.types import BehaviorPrimitive, CompilationError
+from questfoundry_compiler.spec_compiler import SpecCompiler
+from questfoundry_compiler.types import BehaviorPrimitive, CompilationError
 
 __all__ = ["BehaviorPrimitive", "CompilationError", "SpecCompiler"]
